+++ js/main.js
@@ -1,4 +1,9 @@
 import {selectSlide} from "./utils/util.js";
 import screenIntro from './templates/welcome';
 
+import updateGame from './components/renderGameScreen.js';
+import INITIAL_GAME from './data/game-data';
+
 selectSlide(screenIntro.element);
+// debugger;
+// selectSlide(updateGame(INITIAL_GAME));
