+++ js/view/gameScreen-view.js
@@ -1,11 +1,8 @@
 import AbstractView from './abstract-view';
-import {questions, questionTypes, stats} from "../data/game-data";
+import {questions, stats} from "../data/game-data";
 import renderQuestion from "../components/renderQuestions";
 import renderHeader from "../components/header";
 import renderStats from "../components/renderStats";
-import {getAnswer} from "../components/getAnswer";
-
-import GameOneTemplate from '../view/gameOne-view';
 
 export default class GameView extends AbstractView {
   constructor(stat) {
@@ -18,49 +15,17 @@ export default class GameView extends AbstractView {
     const question = questions[currentQuestion];
 
     const GameClass = renderQuestion(question);
-    const gameScreen = new GameClass(question);
+    this.gameScreen = new GameClass(question, this.onAnswer);
 
     return `
     ${renderHeader(lives)}
-    ${gameScreen.template}
+    ${this.gameScreen.template}
     ${renderStats(stats)}
     `;
   }
 
   bind() {
-    const gameTitle = this.element.querySelector(`.game__task`).innerHTML;
-    const form = this.element.querySelector(`form`);
-    const gameAnswers = this.element.querySelectorAll(`input`);
-
-    switch (gameTitle) {
-      case questionTypes.TWO_IMG:
-        const seeGameTwo = (evt) => {
-          if ([...gameAnswers].filter((el) => el.checked).length === 2) {
-            getAnswer(evt, this.element);
-            this.onAnswer();
-          }
-        };
-        form.addEventListener(`change`, seeGameTwo);
-        break;
-
-      case questionTypes.PHOTO_OR_PAINT:
-        const seeGameWide = (evt) => {
-          if ([...gameAnswers].filter((el) => el.checked).length === 1) {
-            getAnswer(evt, this.element);
-            this.onAnswer();
-          }
-        };
-        form.addEventListener(`change`, seeGameWide);
-        break;
-
-      case questionTypes.FIND_PAINT:
-        const selectPic = (evt) => {
-          getAnswer(evt, this.element);
-          this.onAnswer();
-        };
-        form.addEventListener(`click`, selectPic);
-        break;
-    }
+    this.gameScreen.bind(this.element);
   }
 
   onAnswer() {
